+++ sample/compose-passkey-android/src/test/kotlin/dev/webauthn/samples/composepasskey/android/AndroidAppOriginTest.kt
@@ -1,0 +1,16 @@
+package dev.webauthn.samples.composepasskey.android
+
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class AndroidAppOriginTest {
+    @Test
+    fun derives_android_origin_from_signing_certificate() {
+        val origin = androidApkKeyHashOrigin("demo signing certificate".encodeToByteArray())
+
+        assertEquals(
+            "android:apk-key-hash:dq6yglJSBPNx57jffGnnlZMWYfYqLMkq1pQ3HIXibpg",
+            origin,
+        )
+    }
+}